kalman update

+++ quadcopter/dev/include/imu/imu.h
@@ -12,8 +12,8 @@
 #include <gyro.h>
 #include <magnetometer.h>
 
-void imu_get_angles(float angles[]);
-void imu_get_rates(float rate[]);
+void imu_get_angles(double angles[]);
+void imu_get_rates(double rate[]);
 void init_imu(void);
 
 #endif /* IMU_H_ */
